+++ CHANGELOG.md
@@ -132,6 +132,7 @@
 110. Plugin and Mods now have their Names Colorized properly when Loading.
 111. Added HarmonyShield Support for Classes and Structs.
 112. Fixed Internal Failure with mono_free on Non-BleedingEdge Mono Games.
+113. Improved HashCode Implementation.
 
 ---